refactor: move SectionController class functionality to LineController class and remove SectionController class
